feat: move array extensions into separate file
